retarget to netstandard, use CodeGeneration.Roslyn
