+++ src/Amadevus.RecordGenerator.Analyzers/Amadevus.RecordGenerator.Analyzers.csproj
@@ -7,22 +7,10 @@
   </PropertyGroup>
   
   <PropertyGroup>
-    <PackageId>Amadevus.RecordGenerator.Analyzers</PackageId>
-    <PackageVersion>1.0.0.0</PackageVersion>
-    <Authors>Amadeusz Sadowski</Authors>
-    <PackageLicenseUrl>https://github.com/amis92/RecordGenerator/blob/master/LICENSE</PackageLicenseUrl>
-    <PackageProjectUrl>https://github.com/amis92/RecordGenerator</PackageProjectUrl>
-    <PackageIconUrl>https://raw.githubusercontent.com/amis92/RecordGenerator/master/docs/logo.png</PackageIconUrl>
-    <RepositoryUrl>https://github.com/amis92/RecordGenerator</RepositoryUrl>
-    <PackageRequireLicenseAcceptance>false</PackageRequireLicenseAcceptance>
-    <Description>Analyzers and CodeFixes that generate code for [Record]-marked types: constructors and mutators, in separate partial file. See project site for more details.</Description>
-    <PackageReleaseNotes>Bugfix. Doesn't raise false RG0003 'requires regeneration' errors for Records nested in Records. Also now diff is attached to that error message.</PackageReleaseNotes>
-    <Copyright>Copyright © 2018 Amadeusz Sadowski</Copyright>
-    <PackageTags>Amadevus.RecordGenerator.Analyzers, analyzers, codefix, record, constructor, mutator, generator, immutable</PackageTags>
+    <Description>Analyzers and CodeFixes for [Record]-marked types. See project site for more details.</Description>
+    <PackageReleaseNotes>Only left attribute missing diagnostic, removed all others because they were no longer used. See RecordGenerator v0.3.0 release notes.</PackageReleaseNotes>
+    <PackageTags>record; immutable; attribute; generator; amadevus; codegen; analyzer; codefix</PackageTags>
     <NoPackageAnalysis>true</NoPackageAnalysis>
-    <RepositoryType>git</RepositoryType>
-    <Company />
-    <Version>0.3.0</Version>
   </PropertyGroup>
    
   <ItemGroup>
@@ -38,6 +26,10 @@
   <ItemGroup>
     <None Update="tools\*.ps1" CopyToOutputDirectory="Always" Pack="true" PackagePath="" />
     <None Include="$(OutputPath)\$(AssemblyName).dll" Pack="true" PackagePath="analyzers/dotnet/cs" Visible="false" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\Amadevus.RecordGenerator.Attributes\Amadevus.RecordGenerator.Attributes.csproj" />
   </ItemGroup>
 
 </Project>
