+++ src/Amadevus.RecordGenerator.Analyzers/GenerateRecordAttributeDeclarationCodeFixProvider.cs
@@ -10,7 +10,7 @@ using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
 using Microsoft.CodeAnalysis.Formatting;
 
-namespace Amadevus.RecordGenerator
+namespace Amadevus.RecordGenerator.Analyzers
 {
     [ExportCodeFixProvider(LanguageNames.CSharp, Name = nameof(GenerateRecordAttributeDeclarationCodeFixProvider)), Shared]
     public sealed class GenerateRecordAttributeDeclarationCodeFixProvider : CodeFixProvider
@@ -51,16 +51,7 @@ namespace Amadevus.RecordGenerator
         
         private Task<Solution> CreateRecordAttributeDeclarationDocument(Document document, IdentifierNameSyntax declaration, CancellationToken cancellationToken)
         {
-            // get the namespace of document or namespace from attribute usage
-
-            var namespaces = declaration
-                .Ancestors()
-                .OfType<NamespaceDeclarationSyntax>()
-                .Reverse()
-                .Select(nsSyntax => nsSyntax.Name.ToString())
-                .ToArray();
-            var targetNamespace = string.Join(".", namespaces);
-            var text = RecordAttributeDeclarationSource(targetNamespace);
+            var text = RecordAttributeDeclarationSource();
 
             var tree = CSharpSyntaxTree.ParseText(text, cancellationToken: cancellationToken);
             var formattedRoot = Formatter.Format(tree.GetRoot(), document.Project.Solution.Workspace, cancellationToken: cancellationToken);
@@ -68,41 +59,26 @@ namespace Amadevus.RecordGenerator
             return Task.FromResult(doc.Project.Solution);
         }
 
-        internal static string RecordAttributeDeclarationSource(string targetNamespace)
+        internal static string RecordAttributeDeclarationSource()
         {
             var text =
-$@"namespace {targetNamespace}
-{{
-    /// <summary>
-    /// Identifies class or struct that is supposed to have a partial with ctor and mutators generated by source generator.
-    /// </summary>
-    [System.Diagnostics.Conditional(""NEVER"")]
-    [System.AttributeUsage(System.AttributeTargets.Class | System.AttributeTargets.Struct, Inherited = false, AllowMultiple = false)]
-    internal sealed class RecordAttribute : System.Attribute
-    {{
-        /// <summary>
-        /// Gets or sets access modifier of the generated constructor. Valid values: ""public"", ""protected"", ""private"".
-        /// Default is ""public"".
-        /// </summary>
-        public string ConstructorAccess {{ get; set; }} = ""public"";
+@"
+using CodeGeneration.Roslyn;
+using System;
+using System.Diagnostics;
 
-        /// <summary>
-        /// Gets or sets whether mutator methods should be generated (e.g. WithSurname). Default is true.
-        /// </summary>
-        public bool GenerateMutators {{ get; set; }} = true;
-    }}
-}}
+/// <summary>
+/// Identifies class or struct that is supposed to have a partial with ctor and mutators generated by source generator.
+/// </summary>
+[AttributeUsage(AttributeTargets.Class, Inherited = false, AllowMultiple = false)]
+[CodeGenerationAttribute(""Amadevus.RecordGenerator.CodeGenerator"")]
+[Conditional(""CodeGeneration"")]
+public sealed class RecordAttribute : Attribute
+{
+}
 ";
-
-            /* to add sometime in future:
-
-        /// <summary>
-        /// Gets or sets whether collection mutator methods should be generated (e.g. AddItems, ReplaceItems). Default is true.
-        /// </summary>
-        public bool GenerateCollectionMutators {{ get; set; }} = true;
-        
-             */
-            return text;
+            // remove initial newline
+            return text.TrimStart('\r', '\n');
         }
     }
 }
