+++ src/Amadevus.RecordGenerator.Analyzers/Utilities/RecordAttributeProperties.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Amadevus.RecordGenerator.Analyzers
+{
+    internal class RecordAttributeProperties
+    {
+        public const string Filename = "RecordAttribute";
+    }
+}
