+++ src/Amadevus.RecordGenerator.Analyzers/Utilities/RecordPartialProperties.cs
@@ -1,4 +1,4 @@
-﻿namespace Amadevus.RecordGenerator
+﻿namespace Amadevus.RecordGenerator.Analyzers
 {
     internal static class RecordPartialProperties
     {
