+++ src/Amadevus.RecordGenerator/Class1.cs
@@ -1,0 +1,10 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace Amadevus.RecordGenerator
+{
+    class Class1
+    {
+    }
+}
