--- src/Amadevus.RecordGenerator/Concepts/RecordAttribute.cs
@@ -1,28 +1,0 @@
-﻿namespace Amadevus.RecordGenerator
-{
-    /// <summary>
-    /// Identifies class or struct that is supposed to have a partial with ctor and mutators generated by source generator.
-    /// </summary>
-    [System.CodeDom.Compiler.GeneratedCode("RecordGenerator", "0.0.1")]
-    [System.Diagnostics.Conditional("NEVER")]
-    [System.AttributeUsage(System.AttributeTargets.Class | System.AttributeTargets.Struct, Inherited = false, AllowMultiple = false)]
-    internal sealed class RecordAttribute : System.Attribute
-    {
-        /// <summary>
-        /// Gets or sets access modifier of the generated constructor. Valid values: "public", "protected", "private".
-        /// Default is "public".
-        /// </summary>
-        public string ConstructorAccess { get; set; } = "public";
-
-        /// <summary>
-        /// Gets or sets whether mutator methods should be generated (e.g. WithSurname). Default is true.
-        /// </summary>
-        public bool GenerateMutators { get; set; } = true;
-
-
-        /// <summary>
-        /// Gets or sets whether collection mutator methods should be generated (e.g. AddItems, ReplaceItems). Default is true.
-        /// </summary>
-        public bool GenerateCollectionMutators { get; set; } = true;
-    }
-}
