--- src/Amadevus.RecordGenerator/Diagnostic.nuspec
@@ -1,27 +1,0 @@
-﻿<?xml version="1.0"?>
-<package xmlns="http://schemas.microsoft.com/packaging/2011/08/nuspec.xsd">
-  <metadata>
-    <id>Amadevus.RecordGenerator</id>
-    <version>$version$</version>
-    <title>Amadevus.RecordGenerator</title>
-    <authors>Amadeusz Sadowski</authors>
-    <owners>Amadeusz Sadowski</owners>
-    <copyright>Copyright © 2017 Amadeusz Sadowski</copyright>
-    <licenseUrl>https://github.com/amis92/RecordGenerator/blob/master/LICENSE</licenseUrl>
-    <projectUrl>https://github.com/amis92/RecordGenerator</projectUrl>
-    <iconUrl>https://raw.githubusercontent.com/amis92/RecordGenerator/master/docs/logo.png</iconUrl>
-    <requireLicenseAcceptance>false</requireLicenseAcceptance>
-    <description>Analyzers and CodeFixes that generate code for [Record]-marked types: constructors and mutators, in separate partial file. See project site for more details.</description>
-    <releaseNotes>Bugfix. Doesn't raise false RG0003 'requires regeneration' errors for Records nested in Records. Also now diff is attached to that error message.</releaseNotes>
-    <tags>Amadevus RecordGenerator analyzers codefix record constructor mutator generator</tags>
-    <frameworkAssemblies>
-      <frameworkAssembly assemblyName="System" targetFramework="" />
-    </frameworkAssemblies>
-    <developmentDependency>true</developmentDependency>
-  </metadata>
-  <!-- The convention for analyzers is to put language agnostic dlls in analyzers\portable50 and language specific analyzers in either analyzers\portable50\cs or analyzers\portable50\vb -->
-  <files>
-    <file src="*.dll" target="analyzers\dotnet\cs\" exclude="**\Microsoft.CodeAnalysis.*;**\System.Collections.Immutable.*;**\System.Reflection.Metadata.*;**\System.Composition.*;**\System.ValueTuple.*" />
-    <file src="tools\*.ps1" target="tools\" />
-  </files>
-</package>
