--- src/Amadevus.RecordGenerator/Utilities/RecordEntry.cs
@@ -1,19 +1,0 @@
-﻿using Microsoft.CodeAnalysis;
-using Microsoft.CodeAnalysis.CSharp.Syntax;
-
-namespace Amadevus.RecordGenerator
-{
-    [Record]
-    internal class RecordEntry
-    {
-        public RecordEntry(SyntaxToken identifier, TypeSyntax type)
-        {
-            this.Identifier = identifier;
-            this.Type = type;
-        }
-
-        public SyntaxToken Identifier { get; }
-
-        public TypeSyntax Type { get; }
-    }
-}
