--- src/Amadevus.RecordGenerator/packages.config
@@ -1,16 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Microsoft.CodeAnalysis.Analyzers" version="1.1.0" targetFramework="portable45-net45+win8" developmentDependency="true" />
-  <package id="Microsoft.CodeAnalysis.Common" version="1.0.1" targetFramework="portable45-net45+win8" />
-  <package id="Microsoft.CodeAnalysis.CSharp" version="1.0.1" targetFramework="portable45-net45+win8" />
-  <package id="Microsoft.CodeAnalysis.CSharp.Workspaces" version="1.0.1" targetFramework="portable45-net45+win8" />
-  <package id="Microsoft.CodeAnalysis.Workspaces.Common" version="1.0.1" targetFramework="portable45-net45+win8" />
-  <package id="Microsoft.Composition" version="1.0.27" targetFramework="portable45-net45+win8" />
-  <package id="NuGet.CommandLine" version="3.5.0" targetFramework="portable45-net45+win8" developmentDependency="true" />
-  <package id="System.Collections" version="4.3.0" targetFramework="portable45-net45+win8" />
-  <package id="System.Collections.Immutable" version="1.4.0" targetFramework="portable45-net45+win8" />
-  <package id="System.Reflection.Metadata" version="1.0.21" targetFramework="portable45-net45+win8" />
-  <package id="System.Resources.ResourceManager" version="4.3.0" targetFramework="portable45-net45+win8" />
-  <package id="System.Runtime" version="4.3.0" targetFramework="portable45-net45+win8" />
-  <package id="System.ValueTuple" version="4.4.0" targetFramework="portable45-net45+win8" />
-</packages>
