+++ test/Amadevus.RecordGenerator.Analyzers.Test/Helpers/GeneratorCodeFixVerifier.Helper.cs
