+++ test/Amadevus.RecordGenerator.Analyzers.Test/TheoryHelpers/ITheoryDatum.cs
