--- test/Amadevus.RecordGenerator.Test/packages.config
@@ -1,21 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Microsoft.CodeAnalysis.Common" version="1.0.1" targetFramework="net46" />
-  <package id="Microsoft.CodeAnalysis.CSharp" version="1.0.1" targetFramework="net46" />
-  <package id="Microsoft.CodeAnalysis.CSharp.Workspaces" version="1.0.1" targetFramework="net46" />
-  <package id="Microsoft.CodeAnalysis.VisualBasic" version="1.0.1" targetFramework="net46" />
-  <package id="Microsoft.CodeAnalysis.VisualBasic.Workspaces" version="1.0.1" targetFramework="net46" />
-  <package id="Microsoft.CodeAnalysis.Workspaces.Common" version="1.0.1" targetFramework="net46" />
-  <package id="Microsoft.Composition" version="1.0.27" targetFramework="net46" />
-  <package id="System.Collections.Immutable" version="1.4.0" targetFramework="net46" />
-  <package id="System.Reflection.Metadata" version="1.0.21" targetFramework="net46" />
-  <package id="System.ValueTuple" version="4.4.0" targetFramework="net46" />
-  <package id="xunit" version="2.3.1" targetFramework="net46" />
-  <package id="xunit.abstractions" version="2.0.1" targetFramework="net46" />
-  <package id="xunit.analyzers" version="0.8.0" targetFramework="net46" />
-  <package id="xunit.assert" version="2.3.1" targetFramework="net46" />
-  <package id="xunit.core" version="2.3.1" targetFramework="net46" />
-  <package id="xunit.extensibility.core" version="2.3.1" targetFramework="net46" />
-  <package id="xunit.extensibility.execution" version="2.3.1" targetFramework="net46" />
-  <package id="xunit.runner.visualstudio" version="2.3.1" targetFramework="net46" developmentDependency="true" />
-</packages>
